[FEAT] Options menu and save and load game
Created functionality for all game menu options, this includes closing the application, restarting the level, returning to the main menu. The functionality was created to be able to save a game in progress and be able to load it later in the main menu, the game is loaded with its name, name of the player and the chosen character, the objects where they were left to be able to continue playing.